Added CreateStats protocol for generating per SNP statistics + Updated existing protocol and workflow + Add header to chunks file

+++ protocols/ConcatChunks.sh
@@ -17,23 +17,25 @@ length=${#fromChrPos[@]}-1
 #Info files have headers, remove headers and leave first one
 for ((i=0;i<=length;i++))
 do
-
         chunk=${intermediateDir}/chr${chrom}_${fromChrPos[${i}]}-${toChrPos[${i}]}
-        impute2ChunksMerged[${i}]=${chunk}
 
-	echo "Processing: ${chunk}"
+        if [[ -f ${chunk} ]]
+        then
+                impute2ChunksMerged[${i}]=${chunk}
 
-	if [ "${i}" -eq 0 ]
-	then
-		impute2ChunksInfoMerged[${i}]=${chunk}_info
+                echo "Processing: ${chunk}"
 
-	elif [ "${i}" > 0 ]
-	then
-		sed '1d' ${chunk}_info > tmpfile; mv tmpfile ${chunk}_info
-		impute2ChunksInfoMerged[${i}]=${chunk}_info
-	fi
+                if [ "${i}" -eq 0 ]
+                then
+                        impute2ChunksInfoMerged[${i}]=${chunk}_info
+
+                elif [ "${i}" > 0 ]
+                then
+                        sed '1d' ${chunk}_info > tmpfile; mv tmpfile ${chunk}_info
+                        impute2ChunksInfoMerged[${i}]=${chunk}_info
+                fi
+        fi
 done
-
 
 #Concatenate chunks and info files
 cat "${impute2ChunksMerged[@]}" > "${concatChunksFile}"
